+++ ChangeLog.md
@@ -10,6 +10,9 @@
   ([#120](https://github.com/davep/oldnews/pull/120))
 - Changed the default binding for a number of different commands.
   ([#121](https://github.com/davep/oldnews/pull/121))
+- Changed the way that an article is marked as read once it is displayed.
+- BREAKING CHANGE: Removed `mark_read_on_read_timeout` from the application
+  configuration.
 
 ## v0.9.0